Remove header from editor window
