+++ client/src/components/editor/Editor.js
@@ -2,7 +2,7 @@ import React from "react"
 import EditorHome from './EditorHome'
 import { Fragment } from "react"
 import '../layout/Layout.css'
-import Layout from '../Layout';
+import Layout from '../layout/Layout';
 import EditorSideBarData from './EditorSideBarData'
 
 const headerData = [
@@ -20,7 +20,7 @@ const editor =()=>{
 // Add routes to other editor pages just like in App component
     //Need to have header and sidebar wrapped in a fragmment for other editor pages to use
     <Fragment>
-        <Layout sidebarData={EditorSideBarData} headerData={headerData}>
+        <Layout sidebarData={EditorSideBarData}>
             <EditorHome />
         </Layout>
     </Fragment>
